servo: Merge #8972 - M1501: CSS Error Reporting: Final Steps Last Part (from GauriGNaik:expose-css-errors-1); r=jdm
Source-Repo: https://github.com/servo/servo
Source-Revision: a5a7a8318cbb9cc1338e9894aee61221a2fa3f27

UltraBlame original commit: 42a420700beac11d31abbc6ea7c4c3e77c4f3877

+++ servo/components/script_traits/lib.rs
@@ -145,7 +145,7 @@ pub enum ConstellationControlMsg {
         parent: PipelineId
     },
     
-    ReportCSSError(PipelineId, String, usize, usize, String),
+    ReportCSSError(PipelineId, String, u32, u32, String),
 }
